Module string_cache::atom

Reexports

use phf_shared;

use serde::Deserialize;

use serde::Deserializer;

use serde::Serialize;

use serde::Serializer;

use std::ascii::AsciiExt;

use std::borrow::Cow;

use std::cmp::Ordering;

use std::cmp::Ordering::Equal;

use std::fmt;

use std::hash::Hash;

use std::hash::Hasher;

use std::marker::PhantomData;

use std::mem;

use std::ops;

use std::slice;

use std::str;

use std::sync::Mutex;

use std::sync::atomic::AtomicIsize;

use std::sync::atomic::Ordering::SeqCst;

use shared::STATIC_TAG;

use shared::INLINE_TAG;

use shared::DYNAMIC_TAG;

use shared::TAG_MASK;

use shared::MAX_INLINE_LEN;

use shared::STATIC_SHIFT_BITS;

use shared::ENTRY_ALIGNMENT;

use shared::pack_static;

use self::UnpackedAtom::Dynamic;

use self::UnpackedAtom::Inline;

use self::UnpackedAtom::Static;

Structs

Atom

EmptyStaticAtomSet

PhfStrSet

STRING_CACHE

StringCache

StringCacheEntry

Enums

UnpackedAtom

Constants

BUCKET_MASK

NB_BUCKETS

Statics

STRING_CACHE

Traits

StaticAtomSet

Functions

from_packed_dynamic⚠
Used for a fast path in Clone and Drop.

inline_atom_slice

inline_atom_slice_mut

inline_orig_bytes⚠
For as_slice on inline atoms, we need a pointer into the original string contents.

u64_hash_as_u32

Type Definitions

DefaultAtom
Use this if you don’t care about static atoms.