Module string_cache::atom::repr

Details of the atom representation that need to be shared between the macros crate and the run-time library, in order to guarantee consistency.

Reexports

pub use self::UnpackedAtom::{Dynamic, Inline, Static};

Enums

UnpackedAtom

Constants

DYNAMIC_TAG

INLINE_TAG

MAX_INLINE_LEN

STATIC_TAG

Functions

from_packed_dynamic
Used for a fast path in Clone and Drop.

inline_orig_bytes
For as_slice on inline atoms, we need a pointer into the original string contents.

pack_static