Struct string_cache::namespace::Namespace

pub struct Namespace(pub Atom);

An atom that is meant to represent a namespace in the HTML / XML sense. Whether a given string represents a namespace is contextual, so this is a transparent wrapper that will not catch all mistakes.