Module string_cache::static_sets

Structs

EmptyStaticAtomSet

An empty static atom set for when only dynamic strings will be added

PhfStrSet

A string set created using a perfect hash function, specifically Hash, Displace and Compress.

Traits

StaticAtomSet

A static PhfStrSet