Trait string_cache::StaticAtomSet

pub trait StaticAtomSet: Ord {
    fn get() -> &'static PhfStrSet;

    fn empty_string_index() -> u32;
}

A static PhfStrSet

This trait is implemented by static sets of interned strings generated using string_cache_codegen, and EmptyStaticAtomSet for when strings will be added dynamically.

It is used by the methods of Atom to check if a string is present in the static set.

Required methods

fn get() -> &'static PhfStrSet

Get the location of the static string set in the binary.

fn empty_string_index() -> u32

Get the index of the empty string, which is in every set and is used for Atom::default.

Loading content...