Module string_cache::atom

Structs

Atom

BorrowedAtom

RawSlice

STRING_CACHE

StringCache

StringCacheEntry

StringCacheHeap
A token that represents the heap used by the dynamic string cache.

Enums

UnpackedAtom

Constants

BUCKET_MASK

NB_BUCKETS

Statics

STATIC_ATOM_SET

STRING_CACHE

Functions

copy_memory
Copy of std::slice::bytes::copy_memory, which is unstable.

from_packed_dynamic
Used for a fast path in Clone and Drop.

inline_atom_slice

inline_orig_bytes
For as_slice on inline atoms, we need a pointer into the original string contents.

Implementations