Module string_cache::atom

Structs

Atom

EmptyStaticAtomSet

PhfStrSet

STRING_CACHE

StringCache

StringCacheEntry

Enums

UnpackedAtom

Constants

BUCKET_MASK

NB_BUCKETS

Statics

STRING_CACHE

Traits

StaticAtomSet

Functions

from_packed_dynamic⚠
Used for a fast path in Clone and Drop.

inline_atom_slice

inline_atom_slice_mut

inline_orig_bytes⚠
For as_slice on inline atoms, we need a pointer into the original string contents.

u64_hash_as_u32

Type Definitions

DefaultAtom
Use this if you don’t care about static atoms.